+++ src/services/fs/createEmptyFile.js
@@ -4,8 +4,8 @@ import { normalizePathString } from "../../utils/normalizePathString.js";
 
 export const createEmptyFile = async (comandKey, data) => {
   const normalizePath = normalizePathString(comandKey, data);
-
+  console.log(normalizePath);
   fs.open(normalizePath, "ax", (err, _) => {
-    if (err) process.stdout.write("Operation failed!\n");
+    if (err) console.log("Operation failed!\n");
   });
 };
