+++ src/services/fs/operationCd.js
@@ -1,7 +1,7 @@
 import { access, stat } from "node:fs/promises";
 
-import currentPath from "../utils/current-path.js";
-import { normalizePathString } from "../utils/normalizePathString.js";
+import currentPath from "../../utils/current-path.js";
+import { normalizePathString } from "../../utils/normalizePathString.js";
 
 export const operationCd = async (comandKey, data) => {
   const normalizePath = normalizePathString(comandKey, data);
