+++ src/services/fs/operationUp.js
@@ -1,6 +1,6 @@
 import path from "node:path";
 
-import currentPath from "../utils/current-path.js";
+import currentPath from "../../utils/current-path.js";
 
 export const operationUp = async () => {
   const pathNow = currentPath.getPath();
