+++ src/services/index.js
@@ -1,6 +1,3 @@
 export * from "./closeFileManager.js";
-export * from "./operationLs.js";
-export * from "./operationUp.js";
-export * from "./operationCd.js";
 export * from "./fs/index.js";
 export * from "./os/osOperations.js";
