+++ src/utils/index.js
@@ -1,3 +1,2 @@
 export * from "./getUserName.js";
-export * from "./checkCommandLine.js";
 export * from "./normalizePathString.js";
